Fold the particle system into vm_builtins.c
particles.c and its header are gone. The resources live in runner.h next to
the ds_* ones, and every part_* builtin now does its own work instead of
calling a one-line wrapper. Only four entry points stay external, for the
depth list, the frame step and teardown.

Particles also outlive the type they came from, the way the runtime does.
A destroyed type answers part_type_exists with false and can no longer be
spawned from, but it keeps its pool slot until its last particle dies: those
particles age out looking as they did instead of vanishing on the next step,
and the id is never handed to a new type under them. Checked against the
GameMaker runtime running the same bytecode: it keeps such particles alive,
lets them die on schedule, and draws them in their own type's colour.
